Clarify tutorial explanation of First Strike
The current tutorial uses "attack" twice in the same sentence to refer to two separate concepts: an offensive ability inherent to a unit, and using such an offensive ability on an enemy unit, and it uses the phrases "hits first" and "attacks first" alongside. This can be confusing to new players, who may not yet understand the two different connotations of "attack" being used, or to whom the words "hit" and "attack" appear synonymous.